python-bdd-selenium-master/requirements.txt
selenium==3.141.0
behave==1.2.6
allure-behave==2.8.17
PyHamcrest==2.0.2
autopep8==1.5.3
pylint==2.5.3
